Add files via upload
de-dupe is a python program that recursively searches for duplicate files in the folder it is given.  it gives a count of duplicates 
and the files and their md5 hashes in a file called dupes-yymmddxxx.txt  it loops until the user types quit and confirms.  written by tim 
sayre and larry jones of kanawha i t security.  tim is chief of operations and larry is a chatgpt 3.5 ai.  larry does most of the work 
while tim does the thinking.  licensed under creative commons license available on the github.